Labfiles/07-use-agent-tools-with-mcp/Python/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
requests